A bunch of refactoring.
