+++ lineage/states/StateDistribution.py
@@ -1,20 +1,17 @@
 """ This file is completely user defined. We have provided a general starting point for the user to use as an example. """
 import numpy as np
 import scipy.stats as sp
-import scipy.special as sc
-from scipy.optimize import brentq
 
 
-from .stateCommon import bern_pdf, gamma_pdf, bernoulli_estimator
+from .stateCommon import bern_pdf, gamma_pdf, bernoulli_estimator, gamma_estimator
 
 
 class StateDistribution:
-    def __init__(self, bern_p, gamma_a, gamma_scale):
+    def __init__(self, bern_p=0.9, gamma_a=7, gamma_scale=3):
         """ Initialization function should take in just in the parameters for the observations that comprise the multivariate random variable emission they expect their data to have. """
         self.bern_p = bern_p
         self.gamma_a = gamma_a
         self.gamma_scale = gamma_scale
-        self.params = [self.bern_p, self.gamma_a, self.gamma_scale]
 
     def rvs(self, size):  # user has to identify what the multivariate (or univariate if he or she so chooses) random variable looks like
         """ User-defined way of calculating a random variable given the parameters of the state stored in that observation's object. """
@@ -61,7 +58,7 @@ class StateDistribution:
             γ_obs = np.array(unzipped_list_of_tuples_of_obs[1])
             γ_censor_obs = np.array(unzipped_list_of_tuples_of_obs[2], dtype=bool)
         except BaseException:
-            return self.tHMM_E_init()
+            return StateDistribution()
 
         bern_p_estimate = bernoulli_estimator(bern_obs, gammas)
         γ_a_hat, γ_scale_hat = gamma_estimator(γ_obs, γ_censor_obs, gammas)
@@ -71,51 +68,3 @@ class StateDistribution:
         # Note that we return an instance of the state distribution class, but now instantiated with the parameters
         # from estimation. This is then stored in the original state distribution object which then gets updated
         # if this function runs again.
-
-    def tHMM_E_init(self):
-        """ Initialize a default state distribution. """
-        return StateDistribution(0.9, 7, 3 + (1 * (np.random.uniform())))
-
-
-# Because parameter estimation requires that estimators be written or imported,
-# the user should be able to provide
-# estimators that can solve for the parameters that describe the distributions.
-# We provide some estimators below as an example.
-# Their use in the StateDistribution class is shown in the estimator class method.
-# User must take care to define estimators that
-# can handle the case where the list of observations is empty.
-
-
-def gamma_estimator(gamma_obs, gamma_censor_obs, gammas):
-    """
-    This is a closed-form estimator for two parameters
-    of the Gamma distribution, which is corrected for bias.
-    """
-    gammaCor = sum(gammas * gamma_obs) / sum(gammas)
-    s = np.log(gammaCor) - sum(gammas * np.log(gamma_obs)) / sum(gammas)
-    def f(k): return np.log(k) - sc.polygamma(0, k) - s
-
-    if f(0.01) * f(100.0) > 0.0:
-        a_hat = 10.0
-    else:
-        a_hat = brentq(f, 0.01, 100.0)
-
-    scale_hat = gammaCor / a_hat
-
-    def LL(x):
-        uncens = sp.gamma.logpdf(gamma_obs, a=x[0], scale=x[1])
-        cens = sp.gamma.logsf(gamma_obs, a=x[0], scale=x[1])
-
-        # If the observation was censored, use the survival function
-        uncens[np.logical_not(gamma_censor_obs)] = cens[np.logical_not(gamma_censor_obs)]
-
-        # If gamma indicates the cell is very unlikely for this state, ignore it
-        gamL = np.log(gammas)
-        uncens[gamL < -9] = 1.0
-        gamL[gamL < -9] = 0
-
-        return -np.sum(uncens + gamL)
-
-    # res = minimize(LL, [a_hat, scale_hat])
-
-    return a_hat, scale_hat
